move into parser/slides
