+++ backend/src/lib.rs
@@ -1,4 +1,3 @@
 pub mod errors;
-pub mod parse_slides;
 pub mod parser;
 pub mod tangle;
